requirements.txt
anyio==3.7.0
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.2
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-speech==2.20.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.25.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
Pygments==2.15.1
python-telegram-bot==20.3
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6